Implement mem_read()
Instructions that use memory need mem_read() to read memory
addresses. For this, the Memory struct had to implement
get() and set() which allow to read and write on the
memory abstraction. Also, MemoryRegister struct needed to
implement TryInto<usize> so the memory represented as an array
could be indexed by this type.